Binary Modulus
All of these numeric binary operators are basically the same thing. The code for them necessitates 4 paths, for each combination of integer/floating point, and I'm not sure how to simplify that. Dynamic typing could, but then you incur a penalty at runtime for every numeric option, and also make the code pretty non-obvious for a static language. It's a tough question.