Heap access implemented
I'm still not 100% certain on the name table idea, but I think it should do the job. The actual name identifiers should be interesting. They need to be unique, but the same stringy name needs to come up with  the same number in certain contexts. The code generator may just need to remember that name definitions should always produce a unique identifier.