Expanded opcodes to handle array and string types
Arrays can now be appended to, and strings can be added to. Combined with the ability to index and take the current length, I think this should enable all the collection handling we need. The opcode definitions are getting very ugly now, though. This VM is more complex than my earlier VM, and this structure doesn't really seem to be holding.