VM needed a couple of modifications to work how I'd started to expect it to.
Most notably, I'd been using -1 as a shorthand for anonymous variables. The VM had no idea I was doing this, so dutifully crashed. It now does not crash.

The VM also had no idea what a POP_TOP was, which is concerning as it's quite important. It knows now.

The VM had no ability to return from a function without taking a value with it. It does now.