Functions now track their parameter count so that the stack can be properly cleaned up on return
This is actually pretty important, because while the stack does get cleaned up when you leave a function, it isn't cleaned up automatically anywhere else. A function in a loop would eventually eat the entire stack. Functions kinda have to clean up their own parameters, because--save adding an extra instruction just for that--nothing else can. Unlike x86 asm, which stores the function result in EAX, we put it on the stack, meaning that when the return is complete the top of the stack needs to be the function result--not parameters.